// Frontend appbrc/brandcomputers_v1.1/src/app/model/accounting/Provider.ts
export class Provider {
    id: number;
    name: string;
    cif: string;
    tradeRegister: string;
    providerCode: string;
    vatPayer: boolean;
    streetAddress: string;
    city: string;
    county: string;
    country: string;
    iban: string;
    bank: string;
    email: string;
    personOfContact: string;
    phoneNumber: string;
}